Spells and Strengths - 24/08/2023
Added spells(no AOE yet) which all have a certain accuracy which is doubled when playing the arcanist. As well as each weapon has strengths/proficiencies and weaknesses, if the player is of the correct class and the weapon is proficient against the enemy's type they deal double damage. If the weapon is weak to the enemy's type it deals one half damage.